IGNITE-26478 Remove CacheObjectValueContext#kernalContext method
