+++ modules/core/src/main/java/org/apache/ignite/internal/processors/cacheobject/PlatformCacheObjectImpl.java
@@ -19,7 +19,6 @@ package org.apache.ignite.internal.processors.cacheobject;
 
 import org.apache.ignite.internal.processors.cache.CacheObject;
 import org.apache.ignite.internal.processors.cache.CacheObjectImpl;
-import org.apache.ignite.internal.processors.cache.CacheObjectTransformerUtils;
 import org.apache.ignite.internal.processors.cache.CacheObjectValueContext;
 
 /**
@@ -55,7 +54,7 @@ public class PlatformCacheObjectImpl extends CacheObjectImpl {
     private byte[] valueBytesFromArray(CacheObjectValueContext ctx) {
         assert arr != null;
 
-        return CacheObjectTransformerUtils.transformIfNecessary(arr, 0, arr.length, ctx);
+        return ctx.transformIfNecessary(arr, 0, arr.length);
     }
 
     /** {@inheritDoc} */
